Create a JPA ADO implementation for the REST API

+++ src/main/java/com/y4553r/springboot/demo/mycoolapp/rest/EmployeeRestController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -30,19 +31,34 @@ public class EmployeeRestController {
 		return employeeService.findAll();
 	}
 
-	@GetMapping("/employees/{id}")
-	public Employee finById(@PathVariable long id) {
-		return employeeService.findById(id);
+	@GetMapping("/employees/{employeeId}")
+	public Employee getEmployee(@PathVariable long employeeId) {
+		Employee employee = employeeService.findById(employeeId);
+		if(employee == null)
+			throw new RuntimeException("No employee with the provided id was found!");
+		return employee;
 	}
 	
 	@PostMapping("/employees")
-	public void save(@RequestBody Employee employee) {
+	public Employee save(@RequestBody Employee employee) {
+		employee.setId(0);
 		employeeService.save(employee);
+		return employee;
+	}
+	
+	@PutMapping("/employees")
+	public Employee updateEmployee(@RequestBody Employee employee) {
+		employeeService.save(employee);
+		return employee;
 	}
 	
 	@DeleteMapping("/employees/{id}")
-	public void deleteById(@PathVariable long id) {
+	public String deleteById(@PathVariable long id) {
+		Employee employee = employeeService.findById(id);
+		if(employee == null)
+			throw new RuntimeException("No employee with the provided id was found!");
 		employeeService.deleteById(id);
+		return "Deleted employee " + employee.getFirstName() + " " + employee.getLastName() + " with id " + id + ".";
 	}
 
 }
